requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
pytesseract
pdf2image
pillow